+++ packages/patternfly-4/react-core/src/components/Dropdown/DropdownToggle.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { CaretDownIcon } from '@patternfly/react-icons';
 import { Toggle } from './Toggle';
 import styles from '@patternfly/react-styles/css/components/Dropdown/dropdown';
+import { DropdownContext } from './dropdownConstants';
 import { css } from '@patternfly/react-styles';
 
 export interface DropdownToggleProps extends React.HTMLProps<HTMLButtonElement> {
@@ -33,6 +34,8 @@ export interface DropdownToggleProps extends React.HTMLProps<HTMLButtonElement> 
   splitButtonItems?: React.ReactNode[]; 
   /** Accessible label for the dropdown toggle button */
   'aria-label'?: string;
+  /** Accessibility property to indicate correct has popup */
+  ariaHasPopup?: boolean | 'listbox' | 'menu' | 'dialog' | 'grid' | 'listbox' | 'tree';
   /** Type to put on the button */
   type?: 'button' | 'submit' | 'reset';
 }
@@ -51,26 +54,32 @@ export const DropdownToggle: React.FunctionComponent<DropdownToggleProps> = ({
   onToggle = (_isOpen: boolean) => undefined as any,
   iconComponent: IconComponent = CaretDownIcon,
   splitButtonItems,
+  ariaHasPopup,
   ref, // Types of Ref are different for React.FC vs React.Component
   ...props
 }: DropdownToggleProps) => {
   const toggle = (
-    <Toggle 
-      {...props}
-      id={id}
-      className={className}
-      isOpen={isOpen}
-      parentRef={parentRef}
-      isFocused={isFocused}
-      isHovered={isHovered}
-      isActive={isActive}
-      isDisabled={isDisabled}
-      isPlain={isPlain}
-      onToggle={onToggle}
-      {...splitButtonItems && { isSplitButton: true, 'aria-label': props['aria-label'] || 'Select' }}>
-      {children && <span className={IconComponent && css(styles.dropdownToggleText)}>{children}</span>}
-      {IconComponent && <IconComponent className={css(children && styles.dropdownToggleIcon)} />}
-    </Toggle>
+    <DropdownContext.Consumer>
+      {({ toggleTextClass, toggleIconClass }) => (
+        <Toggle
+          {...props}
+          id={id}
+          className={className}
+          isOpen={isOpen}
+          parentRef={parentRef}
+          isFocused={isFocused}
+          isHovered={isHovered}
+          isActive={isActive}
+          isDisabled={isDisabled}
+          isPlain={isPlain}
+          onToggle={onToggle}
+          ariaHasPopup={ariaHasPopup}
+          {...splitButtonItems && { isSplitButton: true, 'aria-label': props['aria-label'] || 'Select' }}>
+          {children && <span className={IconComponent && css(toggleTextClass)}>{children}</span>}
+          {IconComponent && <IconComponent className={css(children && toggleIconClass)} />}
+        </Toggle>
+      )}
+    </DropdownContext.Consumer>
   );
 
   if (splitButtonItems) {