+++ packages/patternfly-3/react-console/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-console@1.12.3...@patternfly/react-console@1.12.4) (2019-08-26)
+
+**Note:** Version bump only for package @patternfly/react-console
+
+
+
+
+
+## [1.12.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-console@1.12.2...@patternfly/react-console@1.12.3) (2019-08-26)
+
+**Note:** Version bump only for package @patternfly/react-console
+
+
+
+
+
 ## [1.12.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-console@1.12.1...@patternfly/react-console@1.12.2) (2019-08-14)