+++ packages/patternfly-4/react-docs/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.28...@patternfly/react-docs@4.10.29) (2019-08-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.10.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.27...@patternfly/react-docs@4.10.28) (2019-08-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.10.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.26...@patternfly/react-docs@4.10.27) (2019-08-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.10.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.25...@patternfly/react-docs@4.10.26) (2019-08-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.10.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.24...@patternfly/react-docs@4.10.25) (2019-08-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [4.10.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.23...@patternfly/react-docs@4.10.24) (2019-08-23)
 
 **Note:** Version bump only for package @patternfly/react-docs