+++ packages/patternfly-4/react-integration/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.21.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@2.21.0...@patternfly/react-integration@2.21.1) (2019-08-26)
+
+**Note:** Version bump only for package @patternfly/react-integration
+
+
+
+
+
 # [2.21.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@2.20.2...@patternfly/react-integration@2.21.0) (2019-08-21)