+++ packages/patternfly-4/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.21.8](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.21.7...demo-app-ts@2.21.8) (2019-08-26)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [2.21.7](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.21.6...demo-app-ts@2.21.7) (2019-08-26)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [2.21.6](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.21.5...demo-app-ts@2.21.6) (2019-08-26)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [2.21.5](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.21.4...demo-app-ts@2.21.5) (2019-08-26)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [2.21.4](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.21.3...demo-app-ts@2.21.4) (2019-08-26)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
 ## [2.21.3](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.21.2...demo-app-ts@2.21.3) (2019-08-23)
 
 **Note:** Version bump only for package demo-app-ts